+++ azure/rs_azure_containerservices/azure_containerservices_plugin.rb
@@ -2,7 +2,7 @@ name 'rs_azure_containerservices'
 type 'plugin'
 rs_ca_ver 20161221
 short_description "Azure Container Services Plugin"
-long_description "Version: 1.0"
+long_description "Version: 1.1"
 package "plugins/rs_azure_containerservices"
 import "sys_log"
 
@@ -88,6 +88,12 @@ plugin "rs_azure_containerservices" do
       type "containerservice"
       path "$href"
       verb "GET"
+    end
+
+    action "update" do
+      type "containerservice"
+      path "$href"
+      verb "PUT"
     end
 
     action "list" do
@@ -189,9 +195,22 @@ define provision_resource(@declaration) return @resource do
   end
 end
 
+define handle_retries($attempts) do
+  if $attempts <= 6
+    sleep(10*to_n($attempts))
+    call sys_log.detail("error:"+$_error["type"] + ": " + $_error["message"])
+    log_error($_error["type"] + ": " + $_error["message"])
+    $_error_behavior = "retry"
+  else
+    raise $_errors
+  end
+end
+
 define delete_resource(@declaration) do
   call start_debugging()
-  sub on_error: skip do
+  $delete_count = 0
+  sub on_error: handle_retries($delete_count) do 
+    $delete_count = $delete_count + 1
     @declaration.destroy()
   end
   call stop_debugging()