+++ azure/rs_azure_compute/changelog.md
@@ -1,5 +1,12 @@
 Compute Plugin changelog
 
+v1.4 (08-16-2017)
+- Adding vmSizes() method to virtualmachine resource
+
+v1.3 (08-16-2017)
+-----------------
+- adding update() methods to virtualmachine resource
+
 v1.2 (08-01-2017)
 -----------------
 - fixing compute extensions to fail properly